Upgrade to Hibernate ORM 6.6.0.CR1 / Search 7.2.0.Alpha2
