+++ extensions/hibernate-orm/runtime/src/main/java/io/quarkus/hibernate/orm/runtime/recording/PrevalidatedQuarkusMetadata.java
@@ -6,6 +6,7 @@ import java.util.Map;
 import java.util.Set;
 import java.util.UUID;
 import java.util.function.Consumer;
+import java.util.function.Supplier;
 
 import org.hibernate.MappingException;
 import org.hibernate.SessionFactory;
@@ -30,6 +31,7 @@ import org.hibernate.mapping.FetchProfile;
 import org.hibernate.mapping.MappedSuperclass;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.Table;
+import org.hibernate.metamodel.mapping.DiscriminatorType;
 import org.hibernate.query.named.NamedObjectRepository;
 import org.hibernate.query.sqm.function.SqmFunctionDescriptor;
 import org.hibernate.query.sqm.function.SqmFunctionRegistry;
@@ -293,6 +295,12 @@ public final class PrevalidatedQuarkusMetadata implements MetadataImplementor {
     @Override
     public Component getGenericComponent(Class<?> componentClass) {
         return metadata.getGenericComponent(componentClass);
+    }
+
+    @Override
+    public DiscriminatorType<?> resolveEmbeddableDiscriminatorType(Class<?> embeddableClass,
+            Supplier<DiscriminatorType<?>> supplier) {
+        return metadata.resolveEmbeddableDiscriminatorType(embeddableClass, supplier);
     }
 
     public Map<String, PersistentClass> getEntityBindingMap() {
